requirements.txt
openai==1.14.2
flask==3.0.2
python-dotenv==1.0.1
pymongo==4.6.2